feat: 관리자 기능 구현

+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,21 +2,27 @@ package roomescape.service;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.auth.AuthorizationException;
+import roomescape.auth.Role;
 import roomescape.controller.dto.ThemeResponse;
 import roomescape.domain.Duration;
 import roomescape.domain.EntityId;
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
+import roomescape.domain.User;
 import roomescape.exception.EntityNotFoundException;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.InUseEntityException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
+import roomescape.repository.UserRepository;
 import roomescape.repository.dto.ReservedTheme;
 import roomescape.service.dto.ThemeCreateCommand;
 import roomescape.service.mapper.ThemeResponseMapper;
@@ -28,12 +34,17 @@ public class ThemeService {
 
     private final ThemeRepository themeRepository;
     private final ReservationRepository reservationRepository;
+    private final UserRepository userRepository;
     private final ThemeResponseMapper themeResponseMapper;
 
     @Transactional
     public ThemeResponse create(
+            EntityId managerId,
             ThemeCreateCommand command
     ) {
+        User user = findUserById(managerId);
+        validateManageAuthority(user, command.shopId());
+
         EntityId id = EntityId.random();
         Theme theme = new Theme(
                 id,
@@ -48,8 +59,8 @@ public class ThemeService {
     }
 
     @Transactional(readOnly = true)
-    public List<ThemeResponse> findAll() {
-        return themeRepository.findAll()
+    public List<ThemeResponse> findByShopId(UUID shopId) {
+        return themeRepository.findByShopId(EntityId.fromUuid(shopId))
                 .stream()
                 .map(themeResponseMapper::map)
                 .toList();
@@ -57,12 +68,16 @@ public class ThemeService {
 
     @Transactional(readOnly = true)
     public List<ReservedTheme> findMostReserved(
+            UUID shopId,
             long limit,
             Duration duration
     ) {
         List<Reservation> reservations = reservationRepository.findBetweenDuration(duration);
         Map<EntityId, Long> themeReservedCounts = collectCountByThemeId(reservations);
-        Map<EntityId, Theme> themes = themeRepository.findByIds(themeReservedCounts.keySet());
+        Map<EntityId, Theme> themes = themeRepository.findByThemeIdsAndShopId(
+                themeReservedCounts.keySet(),
+                EntityId.fromUuid(shopId)
+        );
 
         return themeReservedCounts.entrySet()
                 .stream()
@@ -73,11 +88,36 @@ public class ThemeService {
     }
 
     @Transactional
-    public void delete(EntityId themeId) {
+    public void delete(
+            EntityId managerId,
+            EntityId shopId,
+            EntityId themeId
+    ) {
+        User manager = findUserById(managerId);
+        validateManageAuthority(manager, shopId);
+        Theme theme = findThemeById(themeId);
+        validateThemeOwnership(shopId, theme);
+
         validateThemeNotUsed(themeId);
 
         boolean deleted = themeRepository.delete(themeId);
         validateDeleted(deleted, themeId);
+    }
+
+    private void validateManageAuthority(User manager, EntityId shopId) {
+        if (manager.role() != Role.MANAGER) {
+            throw new AuthorizationException("매니저가 아닙니다.");
+        }
+
+        if (!Objects.equals(manager.managingShopId(), shopId)) {
+            throw new AuthorizationException("해당 매장에 대한 관리 권한이 없습니다.");
+        }
+    }
+
+    private void validateThemeOwnership(EntityId shopId, Theme theme) {
+        if (!shopId.equals(theme.shopId())) {
+            throw new AuthorizationException("해당 테마에 대한 접근 권한이 없습니다.");
+        }
     }
 
     private void validateThemeNotUsed(EntityId themeId) {
@@ -127,5 +167,21 @@ public class ThemeService {
                 theme.imageUrl(),
                 themeReservedCount.getValue()
         );
+    }
+
+    private Theme findThemeById(EntityId themeId) {
+        return themeRepository.findById(themeId)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        ErrorCode.THEME_NOT_FOUND,
+                        "테마를 조회할 수 없습니다. themeId = " + themeId
+                ));
+    }
+
+    private User findUserById(EntityId userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        ErrorCode.USER_NOT_FOUND,
+                        "식별자로 회원을 조회할 수 없습니다."
+                ));
     }
 }
